+++ R/D_regularized.R
@@ -14,7 +14,7 @@
 #' @param fold Logical. Is regularization applied across sample folds with separate predictions for each fold? (Default FALSE)
 #' @param fold.var Character string. Name of the fold variable. (default NULL)
 #' @param pcc Logical. Include probabilities of correct classification? Default FALSE.
-#'
+#' @param auc Logical. Include area under the receiver operating characteristics? Default FALSE.
 #' @return
 #' \item{D}{Multivariate descriptive statistics and differences.}
 #' \item{pred.dat}{A data.frame with predicted values.}
@@ -87,7 +87,8 @@ D_regularized <-
            size = NULL,
            fold = FALSE,
            fold.var = NULL,
-           pcc = FALSE) {
+           pcc = FALSE,
+           auc = FALSE) {
 
     # out-of-bag and folds (fold_out)?
     if (out & fold) {
@@ -102,7 +103,8 @@ D_regularized <-
         rename.output = rename.output,
         size = size,
         fold.var = fold.var,
-        pcc = pcc
+        pcc = pcc,
+        auc = auc
       )
     } # out-of-bag and no folds (out)?
     else if (out & !fold) {
@@ -116,7 +118,8 @@ D_regularized <-
         s = s,
         type.measure = type.measure,
         rename.output = rename.output,
-        pcc = pcc
+        pcc = pcc,
+        auc = auc
       )
     } # not out-of-bag and folds (fold)?
     else if (!out & fold) {