+++ NEWS.md
@@ -5,6 +5,7 @@
 * Fixed joining data frames by fold.var in D_regularized_fold functions
 * Added statistical inference to d_pooled_sd
 * Added probabilities of correct classification option to D_regularized_out and D_regularized_fold_out
+* Added area under the receiver operating characteristics to D_regularized_out and D_regularized_fold_out
 
 # multid 0.1.0